remove unit tests
I think the right approach to testing this, is comparing with the reference
implementation where possible. It is hard to test individual pieces, because
there is a lot of boilerplate to set up when using actual flac files as input.
An other possibility would be to use specially prepared files that represent
e.g. a single metadata block. Comparing output is the easiest solution here I
think, especially because you can automate the comparison and then run it on a
large collection of files, a kind of probablistic approach to testing.