implement mid-side stereo decoding
Though there is a problem with wraparound, that may not be a problem at all,
depending on how the stream is encoded; the specification is not clear on this
point. The reference decoder seems to ignore the problem, maybe the encoder
encodes in such a manner that decoding need not take overflow into account.