upgrade to new IO and ReadExt
This also modifies a number of functions to be generic in the reader, rather
than take a reader object reference. Eventually all functions should follow
this pattern; I prefer compile-time polymorphism over runtime polymorphism. The
reason for the switch at this point is convenience: a `&mut io::Read` in an
immutable slot cannot be passed to a function that expects a `&mut ReadExt`,
even though `ReadExt` is implemented for anything that implements `Read`. This
can be resolved by making the slot mutable, but I think generics are a better
solution.

The tests do not compile yet, some things might have broken.