start moving away from num::Int
The deprication message says it has been replaced by rust-lang/num, but I
dislike having dependencies for this project, I want this to be self-contained.
Additionally, the requirements of rust-lang/num::Num are actually stronger than
what I need (at least, I think so). Furthermore, at this point the library does
not support 64-bit samples (nor does the reference implementation), so it makes
sense to exclude this as an allowed sample type. Having a custom numeric type
might be bad for compatibility with other crates, but at this point I do not
see where such compatibility might be useful; the generics are provided to
allow decoding to narrower integers when the sample width is known, but you can
use i32 in any case just to be safe, and I expect users of this library to use
a concrete type, not a generic. If this turns out to be wrong in the future,
it will not be hard to switch to an other sample type, because essentialy it
just groups a few numeric traits.