implement fixed prediction
This appears to work more or less, but I do not have a sample-by-sample
verification at this point. For one test sample that decodes a fair amount,
there now are less pops, and the result sounds more like the original. Still,
there are some pops, so there must be a mistake. It is not clear to me what the
correct procedure for prediction is. Never overflow? Use i64 and wraparound?
Use sample size and wraparound? Now I think of it, the behaviour for LPC
prediction could be wrong as well. This may very well be the reason that some
test samples still fail with an LPC prediction error, actually.