add associated unsigned type
This type is used for `rice_to_signed`. It feels a bit heavy and all, but it
appears to work. Maybe a more elegant solution could be provided in the future.
The unsigned type is one bit wider, which explains why good FLAC files were
failing to decode: Rice param was maximal, and before this was discarded as too
big. (This demonstrates how good the Rice coding is actually, because for most
files it failed pretty far in the file, so this is really uncommon.)

This marks an important version: for the first time, Claxon managed to decode
an entire FLAC file, and the result was in complete agreement with the
reference decoder. Not only that, _three_ of my test samples decoded properly
before encountering one that failed.