begin moving away from std::num
The `FromPrimitive` trait has been removed in nightly, which is a pain. I
refuse to use the `num` crate (a dependency on `rand` for generic primitive
integers? Why?), so now all relevant functionality has to be implemented
manually. There is an advantage though: I can be much more selective in what to
include. Furthermore, there could be a performance benefit. With
`FromPrimitive`, there were a lot of casts followed by an unwrap. Now a no fail
variant is included, which can avoid the unwrap. I think this avoids a bounds
check, because the cast/unwrap version would need to check for overflow and
panic, even in a release build. So maybe this is a good thing after all.

The `Sample` trait feels a bit ad hoc though, there is room for improvement.