add to_i32 method to Sample
There are a lot of places where it is verified that the `Sample` type is not
wider than 32 bits. This is now an assumption for the type, by having `to_i32`
return an `i32` instead of an `Option<i32>`. FLAC does not support streams with
more than 32 bits per sample, so this assumption is safe.