always decode into a wide buffer
This is based on the following observation: if the stream has x bits per
sample, then by channel decorrelation, the side channel needs x + 1 bits,
and by prediction, the residues could need x + 2 bits to store. So for
decoding, we need an intermediate buffer that is one bit wider anyway, and
possibly two bits wider for the side channel.

The direction I am choosing for now is to work with wider samples everywhere,
and then at the last stage narrow them down to the sample type. This could
be slightly improved upon, by doing left/right/mid/side decoding not in-place,
but combining it with narrowing. Possibly, it could be improved further by
combining prediction, maybe with a kind of stream fusion, such that the wide
residues are used for prediction that is used immediately to compute the final
narrow sample.

There are still a few type inference failures. It might be better to take the
wide sample type as type parameter directly.