implement support for decoding all Rice partitions
Claxon can now fully decode all of the 18 test samples that I used on this
machine, the results agree with the reference decoder. This marks another
milestone for Claxon: I believe it is now able to successfully decode most
real-world FLAC files. (Unencoded binary has not yet been implemented, but
I did not encounter a file that had unencoded binary in it.)