take buffer for reading block as argument
This allows the buffer to be reused to decode multiple frames (this was
the case before as well), but because the caller provides it, there are
no lifetime issues when the returned block is stored, which simplifies a
lot.

Note: benchmarks are now not as efficient as they could be, but this
will be resolved by adding an iterator.