deal with stream end properly
If an EOF happens at a frame boundary that is fine, it is an implicit
end of the stream. When the number of samples was not specified in the
streaminfo metadata block, this is the only way to know that the stream
has ended.

If an EOF happens at a different place than a frame boundary, this is
unexpected, and the appropriate IO error is returned as before.

This also makes the RMS example work.