Bump version to 0.2.1
Also add a documentation link to the crate, now that docs.rs is a thing.

Futhermore, as of this commit I'll capitalize commit messages. I
switched to this style for most of my repositories, and I keep making
mistakes here. A release seems like  an appropriate point to make the
switch.