Benchmark i16 and i32 decoding separately
All the machinery to allow generic decoding into i16 and i32 is quite
complex. If it is no performance win, then perhaps it is not worth
keeping. Before deleting it, measure to make sure that the performance
hist is worth the simpler code.

Results so far seem to indicate that for some files there is virtually
no difference, whereas for other files the hit is substantial.