Add a rudimental benchmark script
More accurate measurements are on the way, but from this we can already
quickly guage what is going on.

I picked five real-world songs in a variety of music styles to benchmark
against. All of these files have been re-encoded with Flac 1.3.1 at the
highest quality level (-8).

Currently these benchmarks seem to suggest: decoding 16-bit integers is
slightly faster, and decoding with Claxon takes about 5 times as long as
decoding with the reference decoder (the reference decoder also writes
its output, which the benchmark program does not because I don't want to
measure IO speed).