Force inline shifts in input reader
Measurements using Perf showed that these were not inlined, and together
responsible for a significant portion of CPU time. Also use the wrapping
shift because it does not panic. (It does not matter in release mode
though.) There should really be a single instruction that can shift by
the integer width, and produce zeroes in the case of integer width. (Not
the "mask the shift value" behavior.)