Use u32 instead of u8 in many places
An 8-bit integer was used in many places because it was sufficient for
the required range. But if it is not going to be stored in a collection,
space is not really an issue, and CPUs can do arithmetic with 32-bit
integers more efficiently than 8-bit integers, so use u32 where
appropriate.