Add a function for reading single bits
This should speed up things a bit, because read_leq_u8 was the hottest
function, and it is reading single bits 99.8% of the time (at least for
one of my test files). Most of these reads are for reading unary encoded
numbers, which can be done in an even faster way. But let's see how far
this gets us.