Shave off two instructions of rice_to_signed
It turns out the compiler was able to find a very efficient way to do
this on its own, although my use of i64 instead of u64 forced it to add
a few more instructions. By using u64 it is possible to get rid of
these. Then encode the optimized version that the compiler found in
Rust, to make sure we get this optimization unconditionally.