Fix LPC predictor order to be 12 always
Other orders less than 12 can all be expressed as an order of 12, by
setting the first few coefficients to 0. This makes the code more
uniform, and hopefully allows the compiler to optimize better, because
it can make assumptions about the number of elements in the slice. A
prelude is needed to handle the first 12 samples in the case of a lower
order.

Oreders up to 32 are allowed by the spec, but orders larger than 12 did
not occur in any of my real-world test files.

The speedup because of this is considerable, almost 20% on my machine.
The generated code is still not vectorized optimally though, so perhaps
further improvement is possible.