Get rid of the wide sample buffer
When assuming subset FLAC (sample width up to 24 bits per sample), there
is no need for a wide buffer. Even with the channel delta, samples are
at most 25 bits, so a 32 bit integer is sufficient everywhere. This gets
rid of the widening step and compacts the data a bit for better
locality. Hence this also speeds up decoding a bit, by about 1 ns per
sample on my machine (about 5%).