Implement custom buffered cursor reader
The main issue with the regular io::Read, is that it does not support
efficient reading of bytes behind an io::Cursor: this involves doing a
memcpy for a single-byte slice, which is very wasteful. It would be
possible to build upon io::BufRead, but that one does not expose its
internal cursor, so making the CRC-computing wrapper on top of that is
impossible. Rolling my own is not such a bad idea anyway, because it can
be specialized for exactly this use case.