Increase hard-coded buffer size
It does not appear to matter a lot for performance in practice, but in
theory this should cut the number of system calls in half when backed
directly by a File. 2048 bytes is still not a lot, so this is probably
fine even in low-memory situations.

The limit should be made configurable in the future though.