Take faster Hound from repository for now
I’ll make a new release of Hound after I polish the API a bit. For now
this should be sufficient. (And now Cargo can finally stop nagging me
about using the "paths" feature which the docs are telling me is exactly
what I need, but Cargo keeps telling me not to use.)