Do not fill buffer with useless zeros initially
This could expose garbage memory to Claxon internally. But then again,
the previous way of dealing with this could also, except in that case
the garbage was always old samples, now it could also be memory left
over from something else.

Fortunately this is safe to do, because the buffer can only be recovered
once a frame has been decoded successfully, and in that case it will
have overwritten all elements in the buffer.