Add fuzzing target for cargo-fuzz
I tried running AFL a long time ago, but it was in a broken state at the
time. Now this cargo-fuzz thing comes along, and I was able to set it up
with a few commands. Awesome! So far it hasn't caught anything. Well, it
did catch a few things, but that was me using it wrong. I'll leave it
running for a while.