Allow wrapping when decoding non-subset files
In debug mode, Rust panics when an arithmetic operation overflows. In
this case, these overflows are safe. They may occur for non-subset
files, and in that case we cannot decode them properly, but the library
does not crash. It just decodes garbage data, which may be detected if
the md5sum is verified.

This issue was identified by libfuzzer using cargo-fuzz.