Use a fold, not sum, for Rust 1.13 compatibility
It turns out, using `iterator::sum` to sum a `Wrapping<i32>` does not
work on Rust 1.13. It does work on later versions, but I want to support
Rust 1.13. So I have to make the iterator chain a bit uglier. At this
point it might be more readable to just use a loop. But now that the
Rust Book took this piece of code as an example, I'd better keep it ...