Add guard for negative qlp shift
The spec does allow this, but it does not happen in practice. As it
would complicate shifting due to arithmetic shift by a negative amount
not being supported, I'll emit an "unsupported" error instead. If
anybody has a real-world FLAC file that fails on this, a specialized
version that deals with this case can be written. For now it is not
worth the effort.