Avoid integer overflow when decoding LPC
This does not occur for valid FLAC files, but it might for invalid ones.
In that case we wrap and produce garbage, however the application should
not crash in debug mode due to Rust panicking on overflow.

This overflow was discovered by libfuzzer and cargo-fuzz.

The performance impact of this change is not significant. My benchmarks
show a larger standard deviation than the precentage improvement.