Speed up reading application metadata block
I used to fill the buffer with zeros before, but they would be
overwritten later anyway. This is not such an issue, except that filling
this buffer with zeros by using an iterator, is apparently terribly
inefficient.

The performance difference in debug mode is as much as a factor 1000.
From ~2.3 seconds to 0.0025 seconds. In release mode this does not
matter so much.

This issue was discovered as a slow case using libfuzzer and cargo-fuzz.