Add two smaller fuzzing targets
One decodes only the header. The idea is that the corpus for this one
can be used as input to the other stages. Only reading the header should
be faster than reading the entire file.

The other target decodes a single block. For large files this could make
sure that no time is wasted on first decoding blocks which are valid
anyway; this should force the fuzzer to discover interesting samples
sooner.