Add short test sample with a length of 4 samples
This sample is currently not handled correctly, and the additional
tests for this sample fail. It triggers an edge case in the frame header
decoder, which reports block sizes smaller than 16 as invalid. But this
is not the case. The block size in the streaminfo must not be smaller
than the block size in the frame header, but a frame itself can be
shorter. In particular, the last block in a stream with a fixed block
size must be shorter, if the stream length is not a multiple of the
block size, which it rarely is. Until now this was not discovered,
because the size of the last block -- although not equal to the fixed
block size -- is rarely smaller than 16 bytes. That happens with a
probability of 15/4096 for a fixed block size of 4096. That is about
one time in 273.