Verify against in-memory files, do not store .wav
The reference decoder normally produces a wav file, which we then read
to compare against. But this clutters the file system with wav files,
and it there are a lot of files to tests, this becomes infeasible. And
it might be unwanted too. So make flac decode to stdout, capture the
result, and decode that as wav. No files are involved in this.