Fix decoding of subframe with wasted bits
It turns out that the wasted bits should be subtracted from the target
bits per sample.

I have one file which was previously decoded incorrectly, which now
decodes properly, identical to the reference decoder. Unfortunately it
is quite a big file, so I cannot add it as a regression tests. I will
try to produce a minimal example to reproduce the bug and add it later.