Add helpful error message for ID3 headers
A FLAC stream that starts with an ID3 header is not a valid stream. The
reference decoder supports it for convenience, decoders are not required
to support it. And if a user wants to support this after all, it can be
done by first reading them, and then building the Claxon reader at the
start of the FLAC header.