Put breaking changes first in changelog
I was reviewing package upgrades today, and I stumbled upon the
changelog of pyOpenSSL [1]. It lists backwards incompatible changes
before anything else, and that is the right thing to do. As a consumer
of a package (perhaps as a third-party dependency), this is pretty much
the only thing I care about. With dozens or even hundreds of
dependencies, reading every changelog carefully becomes unfeasible. From
now on I will list breaking changes first.

[1]: https://pyopenssl.readthedocs.io/en/latest/changelog.html