Put release dates in changelog
They are also available in the Git tags, of course. But having them in
the changelog is very convenient. As a consumer of a library, I am in a
hurry, and I am not going to look around for tags, I just want to
quickly gloss over the changelog, see how outdated my current version
is, and what has changed since.