Implement ReadBytes for io::Cursor
This is not used internally, but it will be useful to external consumers
of ReadBytes. There are none at the moment, but I plan to expose
metadata block reading, which takes a ReadBytes input. For data coming
from the mp4parse crate, a Vec<u8>, wrapping those in an io::Cursor
would be the sensible way to go.

It would be possible to implement ReadBytes for R: Read too, but that
would be inefficient, because Read does not have a backing buffer.
Fortunately, Cursor has.