Add example that decodes flac embedded in mp4
It seems though, that the output is incorrect: the sha256sum of one test
sample decoded directly, and of the wav decoded from mp4 (encoded from
the original flac file with "ffmpeg -acodec flac"), are different.

But for now, I am happy that seeking to some offset in the file,
constructing a FrameReader, and reading frames, actually works!