Guard against malicious Vorbis comment blocks
A maliciously crafted block could contain lengths that are very big.
Because buffers of the right length are allocated immediately, this
could be abused to cause the program to go out of memory, and mount
a denial of service attack in this way.

This is avoided by limiting the size of a Vorbis comment block to 1 MiB,
which should be sufficient for all practical purposes.