Verify Vorbis comment tag names
The restrictions are stricter than being valid UTF-8. Enforcing these
restrictions has the advantage that we don't have to check whether the
tag name is valid UTF-8, because it is implied.

Also get rid of the encoding error variant, users do not care about
this. What matters is that a file that does this is ill-formed, and
recovering part of the string is not something that users would want
to do.