Deal with idiosyncrasies in metaflac output
Sometimes the tests would fail, because a tag contains newlines. There
is no perfect way to test this with metaflac, because it does not escape
such newlines, so we cannot distinguish between end-of-comment, and
a genuine newline. This tries to deal with it a bit by reading the
multiple lines, although it means that information about carriage
returns is lost.

Furthermore, I encountered one file where the vendor string contains
null bytes, and metaflac did not print any output there, likely because
it uses C routines to handle strings. So if the vendor strings starts
with a null byte, assert that the metaflac vendor string is empty.