Remove double buffering from tests
This was added a long time ago, when Claxon did not do buffering
internally. But then buffering was added (to be able to do lookahead for
efficient decoding of unary numbers), and the tests were never updated
to match. Even worse, I copy-pased some code that added the buffer. This
removes the redundant buffers for the tests.