Put a 10 MiB size limit on application blocks
Like with Vorbis comment blocks, this is to mitigate DoS attacks. Now
for the application block, the 10 MiB memory limit is a bit more
arbitrary; the purpose is application-specific, and applications might
have reasons to have 10 MiB application blocks. But if Claxon were to be
used in e.g. a browser, you don't want it to happily try to allocate a
few GiBs of data by feeding it a maliciously crated file. Again there is
a tradeoff here, but I think safety is a good default.

Note that this does not prevent you from having many big application
blocks, but at least those have to be fed and read fully before parsing
continues. It is not possible to make the application allocate much more
memory than the size of the stream.