Store Vorbis comments in "name=value" format
This avoids two copies: rather than producing two strings, we read into
a buffer, and that buffer is later converted into a String, from which
we take slices to build a (name, value) pair. This way no new strings
have to be allocated for the parts. Benchmarking showed that allocating
and copying strings was one of the bottlenecks of reading metadata.