Put Cargo.lock under source control
This is to make builds reproducible and CI more robust. I've had many
builds of my libraries break without changes to the code, because a
transitive dependency released a breaking change under Cargo's dialect
of "semver compatible". This used to not be a problem for Claxon because
it had no dependencies. But now that the examples do have quite a few
dependencies and transitive dependencies, this is becoming a serious
risk. So let's put Cargo.lock under source control and have builds that
do not depend on the current state of crates.io. The trade off is that
CI does not test compatibility with the latest versions of all
dependencies, but I think that is fine.