Read a few more tags in metadata benchmark
Looking up a tag is currently ding a linear search -- and with a case
insensitive comparison too, more expensive than a regular comparision.
But the benchmarks show that this is not an issue at all; times are
dominated by file IO, even with all files in the page cache. So read a
few more tags, to see if this would show up in the benchmark as even a
perent of total execution time, because then it could be worth
optimizing.