Define struct to control metadata reading
This only defines the struct, it does not yet respect it when reading.

The motivation for doing this, is that I am building an application to
index music metadata, and reading only the Vorbis comment block and
skipping e.g. pictures makes a big difference. Embedded cover art can be
megabytes in size, compared to the few kilobytes of tags. When reading
many files from a spinning disk, this makes a significant difference.
Furthemore, reading these blocks wastes space in the page cache. This
can be the difference between the touched parts of my personal
collection fitting in RAM or not. That in turn, is the difference
between indexing that runs in 0.05 seconds (with warm cache), and
indexing in 80 seconds (straight from rotating disk).