Add EmbeddedReader reader
This ReadOptions is getting out of hand, hard-coding some behaviors for
all desired use cases is becoming too complex. I should just make the
metadata block iterator more convenient to use for end users. And the
key insight here is that to avoid allocations, we can instead expose a
reader, that users can decide to use, or ignore.

For Vorbis comments I could use a similar lazy strategy, where the
comments are not even parsed until a consumer asks for them.