Try to unify ReadBytes and io::Read
I introduced ReadBytes, the io::Read::read_exact() method did not exist
yet, and reading integers with a specific endianness had just been
removed. This was during the std::io refactor, in the pre-1.0 days of
Rust. It seems that nowadays there is no benefit to ReadBytes, possibly
apart from the skip() method that could be implemented using a seek once
specialization is a thing. But I think that could happen anyway. So I
think I can drop ReadBytes in the public API at least, and have a
blanket ReadBytes impl for every io::Read. The BufferedReader could
still have a more specialized implementation, where skipping bumps the
pointer. I think this would also work for io::cursor. But that would be
a next step.

This commit replaces ReadBytes with just io::Read in the public
interface in the metadata module. Everything is broken still.