Take next step in unifying io::Read and ReadBytes
At this point, I think we can atually do the following:

 * Make ReadBytes an always-implemented extension of io::Read that just
   provides some helper methods implemented in terms of io::Read.

 * Drop BufferedReader. There was an argument in there about io::BufRead
   not being able to support CRC computations on consume, but I am not
   sure why any more. It is also not entirely relevant, because
   BitReader reads only one byte at a time, and it does not rely on the
   buffer.

So it seems that we can simplify things a lot and get rid of custom
traits in the process to improve compatibility!

There is one issue: implementing skip() requires reading all data to be
skipped over, which is a bit wasteful. At least the BufferedReader could
just bump the pointer. We could also require a BufferedReader as input
and still do this, but it is not as good as io::Seek anyway. Perhaps we
should just wait until specialization is a thing, and have fast skips
for readers that can seek.

Another thing I just thought about is that it might not be that hard to
buffer a few more bytes in the Bitreader after all. There are always two
bytes in the frame footer for the CRC, so we could buffer two additional
bytes, do lzcnt on 24 bits rather than 8, and also do more CRC updates
at once (although I am not sure whether that would be beneficial). The
CRC computations are reversible, so at the end of the frame we can
unreead the two footer bytes.

Anyway, things are still severely broken, but I have a better idea of
how to move forward now.