Remove BufferedReader in favour of io::BufRead
My comments indicated that I needed BufferedReader to be able to do CRC
computations on consume, which the io::BufRead apparently could not do.
I now think it could, but even then, that functionality is not used,
the bitstream reads a single bytes at a time. So just drop the
BufferedReader, and use io::BufRead instead for the convenience
constructor.