Expose read_flac_header(), use from medata reader
This is for convenience (users can quickly check if a file is a flac
file), and for symmetry with Hound. MetadataReader::new() now reads the
header. The idea is to have an optional two-stage construction process,
where first you construct a MetadataReader to read metadata, and then a
FlacReader to read audio data.