Set codegen-units=1, avoid 45% perf regression
Since Rust 1.24.0, the number of codegen units is no longer 1, even for
release builds. This speeds up builds slightly, but it reduces the
quality of the generated code tremendously. To get decent performance
out of Claxon, we need to explicitly pass -C codegen-units=1 as
RUSTFLAGS. Update the benchmarks to do so, and add a note in the readme.

Thanks to est31 for pointing this out.

See also https://github.com/rust-lang/rust/issues/53833.