Zero newly allocated memory in decode buffer
Previously, the buffer was allocated to the right size, but the memory
was not zeroed, because the entire buffer will be overwritten anyway, so
zeroing is a waste. But in the presence of bugs, this can lead
uninitialized memory to be exposed.

I did some careful benchmarking, and the performance impact of zeroing
is negligible for the common case where the buffer can be recycled, and
the block size is stable. That means that only initially the buffer
needs to be resized, and subsequent blocks all reuse the buffer without
any resizing (not in capacity, nor in vec len). Because of that, zeroing
is very uncommon, so its expense is not that important, because it is a
cold code path.

I changed the benchmarks to report the raw decode times per sample
(averaged over a block still), and decoded 5 distinct files each 50
times. This made for a total of 1094220 timing measurements. With R we
can do a t-test to see the impact of zeroing.

In the first case, we compare the typical case: recycling the buffer.

  > t.test(nozero_recycle, zero_recycle)

          Welch Two Sample t-test

  data:  nozero_recycle and zero_recycle
  t = -2.8154, df = 2186400, p-value = 0.004872
  alternative hypothesis: true difference in means is not equal to 0
  95 percent confidence interval:
   -0.006822184 -0.001222073
  sample estimates:
  mean of x mean of y
   14.12020  14.12422

  > mean(zero_recycle) / mean(nozero_recycle)
  [1] 1.000285

Values are decode time per sample in nanoseconds, so on my system Claxon
can do a sample in about 14 ns. As we can see the impact of zeroing is
slight, but significant.

In the second case, we truncate the decode buffer to 10 samples before
decoding a block. This is a worst-case scenario for zeroing, because it
has to zero out practically the entire buffer, and it can't be smart
about requesting zeroed pages from the OS, because the first 10 samples
have to be preserved. Truncating the buffer should not negatively affect
the no-zeroing code, because the capacity is still sufficient.

  > t.test(nozero_truncate, zero_truncate)

          Welch Two Sample t-test

  data:  nozero_truncate and zero_truncate
  t = -23.217, df = 2188400, p-value < 2.2e-16
  alternative hypothesis: true difference in means is not equal to 0
  95 percent confidence interval:
   -0.02758650 -0.02329142
  sample estimates:
  mean of x mean of y
   14.10286  14.12830

  > mean(zero_truncate) / mean(nozero_truncate)
  [1] 1.001804

Here we can see a larger impact of about a 0.2% increase in decode time.
That's somewhat painful, given that many of the optimizations in Claxon
were wins of a similar order of magnitude. However, keep in mind that
this is for an atypical case.

As a sanity check, we should see no significant difference in decode
times between nozero_truncate and nozero_recycle, but that is not the
case, nozero_truncate is significantly faster:

  > t.test(nozero_truncate, nozero_recycle)

          Welch Two Sample t-test

  data:  nozero_truncate and nozero_recycle
  t = -13.487, df = 2036300, p-value < 2.2e-16
  alternative hypothesis: true difference in means is not equal to 0
  95 percent confidence interval:
   -0.01985537 -0.01481676
  sample estimates:
  mean of x mean of y
   14.10286  14.12020

One difference is that in the truncate case, we mostly hit the code path
that does an unsafe Vec::set_len(), whereas the recycling case fails the
check for "buffer.len() < new_len" when the buffer has exactly the right
size already, we call Vec::truncate() instead. Digging into the
Vec::truncate source code reveals that it is implemented as a loop that
drops the values. I have not confirmed if this loop is optimized away
(it should be, because i32's don't need to be dropped), but if it is
not, that might explain the difference.

In any case, I think the overhead of zeroing small enough that we can
afford it.

Thanks to Sergey "Shnatsel" Davidoff for getting this discussion
started.