Add benchmark analysis Python script
After struggling with R for a bit I decided to just give up and use
numpy instead. It is noticeably slower, but the API is consistent and
makes sense. When I know what I want to do, there is a straightforward
way to write it, rather than having to go through Stack Overflow for
half an hour to find out how I can shoehorn this particular variation of
that data frame in the shape that some function expets ... Numpy just
has arrays and shapes. Functions implicitly map over the array, and you
can select over which axis. Perfect.