Print only confidence interval bounds
And the centre of the interval, which is not the 0.5-quantile, only it
looks more natural to people. A confidence interval where the parameter
is at a boundary of the interval -- or worse, outside of the interval --
would be a tough sell (even though there is something to say for it).
The centre of the interval, although having no significance, is at least
in the middle of the interval, and there is reasoning behind the
interval too, so that should be better.

I also added a factor 2 which I can't explain but which does make
everything come out nicely. (My inner physicist at work here ...) It
might be that I missed something in the equations and it is supposed to
be there. It might also be that an exponential distribution just models
the noise so badly that all of my calculations are bogus.