Compute reasonable confidence interval on noise
Using the empirical CDF and then getting a CDF for the min of n
observations from that works nicely. I get a number that looks
reasonable to me, and the numbers are consistent with my measurements.

Furthermore, I don't have to make any assumptions about the distribution
of the noise, which is neat too.