Compute p-value when comparing benches
Ok a p-value rolls out, but I don't trust it, it is 1e-10 for my two
benchmark runs of the same commit. Perhaps one run really was a bit
faster, and the benchmark was really sensitive and registered it. But
then I don't want it to. I have a good estimate of the noise now, and
the difference in means is much less than the median noise even.

Perhaps I should model the differences as the difference between two
noise values, rather than just assuming it's normal and doing a t-test.