Try to fit Erlang noise once more
After collecting many, many, iterations (2100 by now), the noise shape
started to change. I am finally at a point where, if I double the number
of iterations, the reported time per sample does not go down any more.
(Not by more than 0.002 ns, I suspect.) So I am somewhat confident that
I found the real minima for each block, and that the noise distribution
I get out of this is correct. And it looks Erlang again, but not Erlang
with k=1 or 2, but with k=12 or k=13. In fact, with 700 iterations it
looks more like k=12, and the more iterations I add, the higher k
appears to be, which is consistent with me suspecting a low k earlier.

This suggests, that the right way to go about it, is to fit an offset
Erlang distribution to the samples we have for each block. From that I
can get confidence intervals and estimates of the time per sample.