Estimate maximum likelihood time minima
The approach: use the minima for times per sample as the initial guess
for the true time per sample (which we assume is t, and we can observe
t + x, where x follows a gamma distribution with k=12). From these
initial assumptions, we can estimate the scale parameter of the
distribution. With the shape and scale parameter now fixed, we can
compute the maximum likelihood estimates for the offset, per frame. (The
offset of the distribution is the true time per sample t.) With those
new offsets, we update our estimate of the scale parameter, and repeat.

The process converges, at least for the data I tried. The mean of the
offset converges to something that looks reasonable to me, consistent
with the minima that I observed over many many more samples (I am using
a subset of the data for testing now). The estimate for k does not
converge to k exactly however, which could mean that my choice of k is
incorrect. Also, the formula I use to estimate k is only an
approximation, within 1.5% of the true value according to Wikipedia. So
getting back 12.271 after putting in 12 seems plausible too.