Work around temporary Rust distribution breakage
Temporarily disable builds that use the nightly toolchain, as it cannot
be installed at the moment due to [1]. This issue will sort itself out
soon, after which this commit can be reverted.

It's a bit of a coincidence that the Travis cron build happened to run
today, of all days; had it run tomorrow, then I probably would not have
noticed a thing.

[1]: https://github.com/rust-lang/rust/issues/57488