Add test sample for non-subset file
This is a single frame extracted from a larger file that could not be
decoded before implementing support for LPC > 12. We deliberately
include only a single frame to make the original audio unrecognizable.

The header still includes the md5 checksum of the audio data of the full
file, so when decoding with libflac, checksum verification fails. I
should still fix the checksum in the file to make the test pass.