Make test sample decodable after surgery
The test sample was constructed by identifying the start and end byte of
the frame that caused the decode to fail, dd-ing that frame out, then
dd-ing out the header of the file too, up to the end of the STREAMINFO
block, and concatenating those. (I first removed all other metadata
blocks with metaflac to ensure that the last block flag is set correctly
in the STREAMINFO block.)

That results in a file that is decodable by Claxon, but it suffers from
two issues:

 * The MD5 signature of the audio data in the STREAMINFO block is now
   wrong, so libflac fails the decode.

 * The number of samples in the STREAMINFO block is now wrong.

I fixed both of these, by zeroing out the MD5 signature, and by setting
the correct length, in a hex editor. Even better would be to compute the
correct MD5 signature, but frankly I never reverse-engineered what it is
a signature of. The spec says "MD5 signature of the unencoded audio
data", but there is no such thing as "unencoded" audio. For example, in
Claxon you always get i32 samples, but often the bit depth will be 16
bits per sample, and the upper bits are zero. Whether you include those
zero bits matters a lot for the signature! And then still, are samples
of channels interleaved? And in what order? In what endianness? So it is
not clear to me what bytes to take the MD5 of.

I searched for an existing tool that can replace the signature with one
that it computes over the data, but I found no such thing, and flac
itself reencodes the audio data too. So for now, just zeroing the
signature will have to do.